fix: faucet creation
